requirements.txt
fastapi
uvicorn
rasterio
numpy
python-dotenv
openai
pytest
